Move /react/src/pages/** to /react/src/**
