+++ react/src/admin/HidableTextButton.tsx
