+++ react/src/admin/ManagedUploaders.tsx
