+++ react/src/admin/UserDetails.tsx
