+++ react/src/analysis/CancelAnalysisDialog.tsx
