+++ react/src/analysis/ChipStrip.tsx
