+++ react/src/participants/ParticipantInfoDialog.tsx
