+++ react/src/utils/components/DetailSection.tsx
