+++ react/src/utils/components/InfoList.tsx
