+++ react/src/utils/functions.tsx
